Need to use int32_t to hold an UDP or TCP port number
Need to be able to hold source port numbers up to value 65535 -
or -1 to indicate source port is not bound to a certain number.

+++ ESP8266Interface.h
@@ -354,7 +354,7 @@ private:
     // Drivers's socket info
     struct _sock_info {
         bool open;
-        int16_t sport;
+        int32_t sport;
     };
     struct _sock_info _sock_i[ESP8266_SOCKET_COUNT];
